pyproject.toml
[project]
name = "prompt-injection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
]
